+++ src/Indice.Hosting/Tasks/QueueOptions.cs
@@ -21,11 +21,11 @@ namespace Indice.Hosting.Tasks
         /// <summary>
         /// Specifies the time interval between two attempts to dequeue new items. Defaults to 300 milliseconds.
         /// </summary>
-        public int PollingInterval { get; set; } = 300;
+        public double PollingInterval { get; set; } = 300;
         /// <summary>
         /// Specifies the maximum time interval between two attempts to dequeue new items. Used as a backoff strategy threshold. Defaults to 5000 milliseconds.
         /// </summary>
-        public int MaxPollingInterval { get; set; } = 5000;
+        public double MaxPollingInterval { get; set; } = 5000;
         /// <summary>
         /// Specifies the time interval between two attempts to cleanup items. Defaults to 0 seconds. Zero meaning off/not enabled.
         /// </summary>
